WIP merge fixes
