+++ include/utils.hpp
@@ -25,3 +25,6 @@ namespace log {
         std::cerr << '\n';
     }
 };
+
+template<class... Ts> struct overloaded : Ts... { using Ts::operator()...; };
+template<class... Ts> overloaded(Ts...) -> overloaded<Ts...>;
